chore: update swagger
