+++ docs/docs.go
@@ -6327,6 +6327,44 @@ const docTemplate = `{
                 }
             }
         },
+        "model.AvailableCMD": {
+            "type": "object",
+            "properties": {
+                "application_id": {
+                    "type": "string"
+                },
+                "default_member_permissions": {
+                    "type": "integer"
+                },
+                "description": {
+                    "type": "string"
+                },
+                "description_localizations": {
+                    "type": "string"
+                },
+                "guild_id": {
+                    "type": "string"
+                },
+                "id": {
+                    "type": "string"
+                },
+                "name": {
+                    "type": "string"
+                },
+                "name_localizations": {
+                    "type": "string"
+                },
+                "nsfw": {
+                    "type": "boolean"
+                },
+                "type": {
+                    "type": "integer"
+                },
+                "version": {
+                    "type": "string"
+                }
+            }
+        },
         "model.Chain": {
             "type": "object",
             "properties": {
@@ -7776,44 +7814,6 @@ const docTemplate = `{
                 }
             }
         },
-        "request.AvailableCMD": {
-            "type": "object",
-            "properties": {
-                "application_id": {
-                    "type": "string"
-                },
-                "default_member_permissions": {
-                    "type": "integer"
-                },
-                "description": {
-                    "type": "string"
-                },
-                "description_localizations": {
-                    "type": "string"
-                },
-                "guild_id": {
-                    "type": "string"
-                },
-                "id": {
-                    "type": "string"
-                },
-                "name": {
-                    "type": "string"
-                },
-                "name_localizations": {
-                    "type": "string"
-                },
-                "nsfw": {
-                    "type": "boolean"
-                },
-                "type": {
-                    "type": "integer"
-                },
-                "version": {
-                    "type": "string"
-                }
-            }
-        },
         "request.ClaimQuestsRewardsRequest": {
             "type": "object",
             "properties": {
@@ -8629,7 +8629,7 @@ const docTemplate = `{
                 "available_cmds": {
                     "type": "array",
                     "items": {
-                        "$ref": "#/definitions/request.AvailableCMD"
+                        "$ref": "#/definitions/model.AvailableCMD"
                     }
                 },
                 "global_xp": {
@@ -10050,6 +10050,12 @@ const docTemplate = `{
                 },
                 "alias": {
                     "type": "string"
+                },
+                "available_cmds": {
+                    "type": "array",
+                    "items": {
+                        "$ref": "#/definitions/model.AvailableCMD"
+                    }
                 },
                 "bot_scopes": {
                     "type": "array",
