+++ pkg/model/discord_guild.go
@@ -2,6 +2,20 @@ package model
 
 import "time"
 
+type AvailableCMD struct {
+	ID                       string `json:"id"`
+	ApplicationID            string `json:"application_id"`
+	Version                  string `json:"version"`
+	DefaultMemberPermissions *int   `json:"default_member_permissions"`
+	Type                     int    `json:"type"`
+	Name                     string `json:"name"`
+	NameLocalizations        string `json:"name_localizations"`
+	Description              string `json:"description"`
+	DescriptionLocalizations string `json:"description_localizations"`
+	GuildID                  string `json:"guild_id"`
+	NSFW                     bool   `json:"nsfw"`
+}
+
 type DiscordGuild struct {
 	ID            string          `json:"id"`
 	Name          string          `json:"name"`
