+++ pkg/request/guild.go
@@ -1,6 +1,10 @@
 package request
 
-import "time"
+import (
+	"time"
+
+	"github.com/defipod/mochi/pkg/model"
+)
 
 type CreateGuildRequest struct {
 	ID       string    `json:"id"`
@@ -8,26 +12,12 @@ type CreateGuildRequest struct {
 	JoinedAt time.Time `json:"-"`
 }
 
-type AvailableCMD struct {
-	ID                       string `json:"id"`
-	ApplicationID            string `json:"application_id"`
-	Version                  string `json:"version"`
-	DefaultMemberPermissions *int   `json:"default_member_permissions"`
-	Type                     int    `json:"type"`
-	Name                     string `json:"name"`
-	NameLocalizations        string `json:"name_localizations"`
-	Description              string `json:"description"`
-	DescriptionLocalizations string `json:"description_localizations"`
-	GuildID                  string `json:"guild_id"`
-	NSFW                     bool   `json:"nsfw"`
-}
-
 type UpdateGuildRequest struct {
-	GlobalXP      *bool           `json:"global_xp"`
-	LogChannel    *string         `json:"log_channel"`
-	Active        *bool           `json:"active"`
-	LeftAt        *time.Time      `json:"left_at"`
-	AvailableCMDs *[]AvailableCMD `json:"available_cmds"`
+	GlobalXP      *bool                 `json:"global_xp"`
+	LogChannel    *string               `json:"log_channel"`
+	Active        *bool                 `json:"active"`
+	LeftAt        *time.Time            `json:"left_at"`
+	AvailableCMDs *[]model.AvailableCMD `json:"available_cmds"`
 }
 
 type HandleGuildDeleteRequest struct {
