+++ pkg/response/guild.go
@@ -11,15 +11,16 @@ type GetGuildsResponse struct {
 }
 
 type GetGuildResponse struct {
-	ID           string   `json:"id"`
-	Name         string   `json:"name"`
-	BotScopes    []string `json:"bot_scopes"`
-	Alias        string   `json:"alias"`
-	LogChannel   string   `json:"log_channel"`
-	LogChannelID string   `json:"log_channel_id"`
-	GlobalXP     bool     `json:"global_xp"`
-	Active       bool     `json:"active"`
-	Icon         string   `json:"icon"`
+	ID            string                `json:"id"`
+	Name          string                `json:"name"`
+	BotScopes     []string              `json:"bot_scopes"`
+	Alias         string                `json:"alias"`
+	LogChannel    string                `json:"log_channel"`
+	LogChannelID  string                `json:"log_channel_id"`
+	GlobalXP      bool                  `json:"global_xp"`
+	Active        bool                  `json:"active"`
+	Icon          string                `json:"icon"`
+	AvailableCMDs *[]model.AvailableCMD `json:"available_cmds"`
 }
 
 type ListAllCustomTokenResponse struct {
